pyproject.toml
[project]
name = "kalorie"
version = "0.0.1"
description = ""
authors = [
    {name = "heko", email = "91366057+hkohko@users.noreply.github.com"},
]
dependencies = [
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
    "pre-commit>=3.5.0",
    "mypy>=1.6.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
